Telnet needs to use config
